LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/import-attributes/invalid-type-attribute-error-expected.txt

CONSOLE MESSAGE: TypeError: Import attribute type "notARealType" is not valid
CONSOLE MESSAGE: TypeError: Import attribute type "" is not valid
CONSOLE MESSAGE: TypeError: Import attribute type "js" is not valid
CONSOLE MESSAGE: TypeError: Import attribute type "javascript" is not valid

PASS Test that invalid module type attribute leads to TypeError on window.
